Move DTO together with other tada-wrapping code
