+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/serialization/KryoRegistrator.scala
@@ -36,9 +36,9 @@ class KryoRegistrator extends BaseKryoRegistrator {
     kryo.register(classOf[facebook4j.Comment])
     kryo.register(classOf[facebook4j.Application])
     kryo.register(classOf[facebook4j.Targeting])
-    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.tadaweb.dto.TadawebEvent])
-    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.tadaweb.dto.TadawebCity])
-    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.tadaweb.dto.TadawebTada])
+    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.streamwrappers.tadaweb.TadawebEvent])
+    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.streamwrappers.tadaweb.TadawebCity])
+    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.streamwrappers.tadaweb.TadawebTada])
     kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.streamwrappers.radio.RadioTranscription])
     kryo.register(classOf[com.github.catalystcode.fortis.spark.streaming.instagram.dto.InstagramItem])
     kryo.register(classOf[com.github.catalystcode.fortis.spark.streaming.instagram.dto.InstagramUser])
