+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/streamwrappers/tadaweb/TadawebAdapter.scala
@@ -1,6 +1,5 @@
 package com.microsoft.partnercatalyst.fortis.spark.streamwrappers.tadaweb
 
-import com.microsoft.partnercatalyst.fortis.spark.tadaweb.dto.TadawebEvent
 import net.liftweb.json
 
 import scala.util.Try
